django/requirements.txt
django-rest-framework==0.1.0
gunicorn==21.2.0
whitenoise==6.6.0
django-livesync==0.5
django-extensions==3.2.3
django-environ==0.11.2
Werkzeug==3.0.1
django-watchman==1.3.0
psycopg2==2.9.9
django-cors-headers==4.3.1
requests==2.31.0
django-allauth==0.58.2
dj-rest-auth==5.0.2
djangorestframework-simplejwt==5.3.1